requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
redis==5.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aioredis==2.0.1
httpx==0.25.2